Set Up Knockback code
Set up the knockback code but still need to figure out why karts aren't able to interact with each other. Edited box colliders on Player GoKart and the NPCKart and marked the "Provides Contacts" on their colliders as true on both as well along with turning off the enemy follow script on the NPC for testing purposes.
(Should be noted that the ball itself can be seen bouncing off the NPCKart's collider so it IS there.)